pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "ai-watermark-recovery"
version = "0.1.0"
description = "AI Watermark Recovery – image processing service with FastAPI, Celery & Redis"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.11"
authors = [{ name = "Hell-Phoenix" }]

dependencies = [
    # --- Web framework ---
    "fastapi>=0.111,<1",
    "uvicorn[standard]>=0.30,<1",
    "python-multipart>=0.0.9",

    # --- Database ---
    "sqlalchemy[asyncio]>=2.0,<3",
    "alembic>=1.13,<2",
    "asyncpg>=0.29,<1",          # async PostgreSQL driver
    "psycopg2-binary>=2.9,<3",   # sync fallback / Alembic migrations

    # --- Task queue ---
    "celery[redis]>=5.4,<6",
    "redis>=5.0,<6",

    # --- Image processing ---
    "Pillow>=10.0,<11",
    "numpy>=1.26,<3",

    # --- ML (PyTorch installed separately via pip+cu*) ---
    "torch>=2.2",
    "torchvision>=0.17",

    # --- Auth / Security ---
    "python-jose[cryptography]>=3.3,<4",
    "passlib[bcrypt]>=1.7,<2",
    "bcrypt>=4.1,<5",

    # --- Settings & utilities ---
    "pydantic>=2.7,<3",
    "pydantic-settings>=2.3,<3",
    "python-dotenv>=1.0,<2",
    "httpx>=0.27,<1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8,<9",
    "pytest-asyncio>=0.23,<1",
    "pytest-cov>=5,<6",
    "httpx>=0.27,<1",
    "ruff>=0.5,<1",
    "mypy>=1.10,<2",
    "pre-commit>=3.7,<4",
]

[project.scripts]
serve = "uvicorn backend.app.main:app --reload"

# ---------- Tool configs ----------

[tool.ruff]
target-version = "py311"
line-length = 100
src = ["backend"]

[tool.ruff.lint]
select = ["E", "F", "I", "N", "UP", "B", "SIM", "RUF"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.mypy]
python_version = "3.11"
strict = true
plugins = ["pydantic.mypy"]
